services/executor/requirements.txt
ccxt==4.2.42
redis==5.0.1
python-dotenv==1.0.1